backend/requirements.txt
strands-agents
boto3
opensearch-py
sympy
strands-agents